2006-05-22  Zoltan Varga  <vargaz@gmail.com>
	* Thread.cs: Change thread_id to int64 to match the unmanaged structure.

svn path=/branches/mono-1-1-7/mcs/; revision=60968

+++ mcs/class/corlib/System.Threading/Thread.cs
@@ -62,7 +62,7 @@ namespace System.Threading
 		private object abort_exc;
 		internal object abort_state;
 		/* thread_id is only accessed from unmanaged code */
-		private int thread_id;
+		private Int64 thread_id;
 		
 		/* start_notify is used by the runtime to signal that Start()
 		 * is ok to return
